Reduce min accuracy of yaw before triggering recalibration
The yaw would only be recalibrated if the error was above 1 deg for 10 ticks. This is quite a large error margin, as we aim for the robots to be within 1.26 deg of its target. Thus, the yaw was reduced by a factor of 4. This was found to increase the accuracy of the robots, while not requiring a very large amount of additional recalibrations.