Change intersectsCircleWithLineSegment to not need circle as argument
Previously, this function was weird to call, as it is a member function that needs the object it is called on as an argument. Either this function should be static, or get the info from the this pointer instead of an argument. In line with other functions in circle, it was decided to do the latter.