Use TeamColor instead of bools and remove getWorldBot
Using bools for the teamcolor is a bit annoying, and can result in confusion as true could mean yellow here, but blue elsewhere. Besides that, the utilities had a getWorldRobot, which was dependent on proto::World. I moved this to RobotHub to keep utilities.h independent from proto files, and because I intend to get rid of this function entirely.