Fix CMake, update utility files
The CMakeLists.txt still set the variable for using CMake modules in the set folder. This is unused, so I removed it.

I also added a few utility functions and fixed the libusb endpoint direction function, that originally was incorrectly the other way around (in was out and out was in)