Fix BasestationMessage, remove channel changing
The BasestationMessage object had some errors with regards to allocating memory. This is fixed now with a set-size buffer size. Downside is that now arrays have to me copied with 'memcpy' for example, but this is more safe to use at least. On top of that, I removed pieces of code that allow for changing the channel of basestations, as this was not working yet. I will implement that later, after the Spain deadline