Fix signedness and default value
Using unsigned int, even though we should only use positive robot ids, only limits us, while there is no downside to just using int (Perhaps for testing purposes). On top of that, it is probably better to not have a standard value of -1 to prevent confusion. Default is always 0 or false