Add scripts for testing
The system tests were actually just scripts for manully testing. I moved these files to the scipts folder. I improved the RobotControl script to have as much freedom in robot commands as RobotHub can provide. I added RobotMonitor that keeps track of which robot has the ball and continuously prints that, nice for testing the ballsensor.