changed automatedSimulation file to bin and made build script scale to available cores
amount of cores used for building was hard coded to 9 cores, which poses a problem when you have less than 9 cores available, for example when using a quad-core processor. Fixed this by scaling it to the amaount of available cores. Also the name 'bin' is more intuitive than 'automatedSimulation', so I changed the name of this file. 'RunSimulation' should now be run from the roboteam_suite file instead, for example: '$ ./bin/RunSimulation.sh'.