Fixed the onlyDecelerateOtherWay test in BBTrajectory1DTest.cpp
The test was wrongly assuming that t=0 and t=1.5 should be in the last part of the trajectory. This is not the case as a trajectory consists of 3 parts, so the last part should be between t=2 and t=3. The test now checks if t=0 is outside the last part and if t=2.5 is inside the last part.