Fixed GoToPos skill and GetBall tactic.
The tactic GetBall was not returning success because the skill GoToPos only checked if the robot was in the right place and not wether it had the ball. Since the robot would (allmost) never be able to get to the right position, as this would mean that the robot was inside the ball, the skill never returned success. Previously this was done by forcing the tactic to be success whenever it had the ball but this would be done outside the state machine, so it was very prone to segfaults. The skill GoToPos now checks whether the robot has the ball as a valid success factor, since the robot would not get the ball if it doesn't need to anyway. Might move this check to the tactic itself at some point.