Fixed ball placer not driving with ball.
The skill GoToPos returns success whenever a friendly robot has the ball. However this is a rare case when we do no want the skill to end when the robot has the ball. This is fixed by excluding the ball placer role from the has ball check in this skill.